feat!: WIP changes to show usage of DetailsView
these changes get us to the point of rendering the Page models
as defined in the fixtures/database. notice that some additional
fields were needed in the models, and the templates needed a lot of
updates.

BREAKING CHANGE: because this is a proof-of-concept/WIP, this commit
has the app in a broken state. Notice that the session is
not being updated.