Updated initial agency launch date for MST
As part of the effort to backfill enrollment data from the early days of Benefits when we used the DMV for eligibility verification, @thekaveman did some research to determine the date of our official deployment of Benefits for MST. We're using [this release](https://github.com/cal-itp/benefits/tree/2021.12.0) to denote launch. I updated the documentation accordingly.