fix(test): remove test that has been revealed to not be meaningful
when the fixture was a Mock object, the test saw `uses_auth_verification`
as a function, but now that the fixture is using an actual model object,
the `@property` decorator obscures the function and makes it look like
a `bool` type.

the intent of the test was to ensure the function is called, but that is
already covered by the `bool` being `True` which is already a part of
other tests.