fix: add volume mapping for client service
the client container needs DJANGO_STORAGE_DIR to be
`/home/calitp/app/data` because the nginx.conf location blocks use that
base path to serve media files.

the volume mapping is needed so that the container has access to the
django.db file and the uploads/ directory from the host side of the
mapping.