refactor: load Bootstrap CSS asynchronously
use approach described in https://www.filamentgroup.com/lab/load-css-simpler/#the-code
where `media` attribute is initially set to `print` so that the browser
will download this resource without blocking page rendering.

then the `onload` event handler sets the `media` attribute to `all` so
that the CSS file is actually used.

our content-security policy has to be modified to allow this event handler
to execute, so `unsafe-hashes` and the hash of the Javascript was added
to the `script-src` setting. the `nonce` approach does not work with
inline event handlers.

you can use this command to get the hash:
```
echo -n "your Javascript here" | openssl dgst -sha256 -binary | openssl base64
```