feat: implement back-end form validation for in-person eligibility
previously we could rely on Django's default handling of the single
`BooleanField`, but since we have multiple checkboxes now, and we don't
want to require all of them, we need to implement our own logic to
require that the checkbox that matches the selected flow was checked.

also added Javascript to the template to show the checkbox for the case
where the form didn't pass back-end validation and therefore we show
the form again with what the user had selected.