refactor: update calls to ClaimsProvider.client_id
note that flow.oauth_config.client_id is wrapped in str to get the
string UUID of the client_id. Not wrapping it in str will result in
getting a UUID object. We may want to add a managed property to the
IdentityGatewayConfig model to return the client_id's UUID string.