refactor: delete unneeded LoginRequired middleware
previously, the main purpose of this middleware was to guard the
`confirm` view. it also guarded the `verified` view, but this was more
just an extension of its main purpose; it's not possible for the user
to reach the `verified` view directly.

since we removed all claims-verification logic from the `confirm` view
in 56ff1a1f and e3011b24, this middleware isn't useful any more and can
be removed.