fix: validation when saving a `TransitAgency` with Switchio config
when using the admin interface to add a new `SwitchioConfig` to a
`TransitAgency`, the Switchio config's `clean` method won't see any
agencies in `self.transitagency_set`. for some reason, this is not the
behavior when doing the same thing in code, so our unit test did not
catch this. I found it by testing manually.

add an `agency` parameter to `clean` so that the agency can be passed
in when it is the one calling `clean`. this gives us the validation we
want (user cannot assign a `SwitchioConfig` with empty fields to an
active agency).