chore: temporary changes to allow the redirect to get further
upon submission of the Switchio card details form, Switchio tries to
redirect back to the URL we gave as `eshopRedirectUri`, but our CSP
settings were blocking it.

these CSP changes allow the redirection attempt to go further, but it
still fails, this time being blocked by Django's CSRF middleware. this
is because the mode we told Switchio to redirect back with is FORM_POST,
and the form doesn't have a CSRF token (because it's being created by
Switchio).