fix(in-person): update system error page
given the in_person templates folder organization, the system_error
template should extend the error-base template that is up one level.
with the default color (grey) the button looks like it is disabled.
changing it to white makes it look like an active button, just like the
rest of the buttons in the in-person app.