docs: Add new card properties to list of what's collected
Also simplifies links to filtered Amplitude event documentation. Theses links still work when clicking them in the public docs, but for some reason launching them from the source in VS Code did not work for me (I was redirected to the unfiltered page), which led me to discover that there is a simpler format that works fine now (and that the existing links also redirect to).