refactor(enrollment): extract origin url handling
we want to reuse this system error view for the In-person error
view, so we extract the origin url handling into a method that can be
overridden. Ideally we would have just used a class attribute,
but since the origin url is dynamic (based on the agency),
we need a method.