chore: remove TransitAgency.staff_group and automatic group creation
given that we are not using the staff_group field, and we
will not use it in the foreseeable future, remove it from the model.
Also remove the automatic creation of the associated group.
Now we are only making the distinction between users that
can access the admin interface (superusers and those in the
settings.STAFF_GROUP_NAME group) and those that can access the in-person
enrollment interface.