feat(admin): improve TransitAgency group validation
by adding another check on the clean and save_model methods, we
ensure that any changes to the short_name are propagated to the
name of the customer_service_group. This prevents inconsistencies
where the short_name and group name do not align, as well as
having dangling groups if an agency's short_name changes.